+++ src/constants/index.js
@@ -5,3 +5,6 @@ export { default as ssoErrorCodes } from './ssoErrorCodes';
 export { default as defaultErrors } from './defaultErrors';
 export { default as packageName } from './packageName';
 export { default as delimiters } from './delimiters';
+
+export { default as CHANNELS } from './channels';
+export { default as EVENTS } from './events';