+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 * Correctly parse `.../_self` permissions object. Refs STCOR-813.
 * Export `getEventHandler` to be able to create events in other modules. Refs STCOR-770.
 * Simplify logout workflow to bypass keycloak confirmation page. Refs STCOR-803.
+* Auto-logout on session expiration, show a "Keep working?" modal first. Refs STCOR-776.
 * After login, only check SSO endpoints when `login-saml` interface is present. Refs STCOR-816.
 
 ## [10.0.0](https://github.com/folio-org/stripes-core/tree/v10.0.0) (2023-10-11)